# RDISCROWD-4906 Optimize stats_dates method used in project_by_shortname