*Issue number of the reported bug or feature request: RDISCROWD-4906*

**Describe your changes**
The original complicated multiple JOIN SQL is really to select all tasks with at least 1 task run and its last finish time. So we can simply get this info from task_run table.

The runtime upon Prod size data(10 million records in task_run) is ~1.3s. The main reason it is still quite slow is because the type of "finish_time" column is text, not timestamp. Thus applying the a function to a column makes index not working, not mentioning there is no index for column "finish_time". Adding an index and changing the column type will be future work. We leave it for now to keep 1.3s query time and utilize cache to compensate the slow query.

**Testing performed**
Local and public QA

**Performance**
Before: 31.3 seconds
After: 1.3 seconds