@kbecker42 2022-02-01
Could we include a before and after of runtime for this query? For example:

_Before: 31.3 seconds
After: 1.3 seconds_